x/adverts/types/messages_advert_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"remap/testutil/sample"
)

func TestMsgCreateAdvert_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateAdvert
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateAdvert{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateAdvert{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateAdvert_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateAdvert
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateAdvert{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateAdvert{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteAdvert_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteAdvert
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteAdvert{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteAdvert{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
